Split DataBlick into pieces.
