+++ assets/js/components/DataBlock/Sparkline.js
@@ -1,0 +1,52 @@
+/**
+ * Sparkline component.
+ *
+ * Site Kit by Google, Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * External dependencies
+ */
+import PropTypes from 'prop-types';
+
+/**
+ * WordPress dependencies
+ */
+import { cloneElement } from '@wordpress/element';
+
+const Sparkline = ( { sparkline, invertChangeColor } ) => {
+	// The `sparkline` prop is passed as a component, but if `invertChangeColor`
+	// is set, we should pass that to `<Sparkline>`. In that case, we clone
+	// the element and add the prop.
+	let sparklineComponent = sparkline;
+	if ( sparklineComponent && invertChangeColor ) {
+		sparklineComponent = cloneElement( sparkline, {
+			invertChangeColor,
+		} );
+	}
+
+	return (
+		<div className="googlesitekit-data-block__sparkline">
+			{ sparklineComponent }
+		</div>
+	);
+};
+
+Sparkline.propTypes = {
+	sparkline: PropTypes.element,
+	invertChangeColor: PropTypes.bool,
+};
+
+export default Sparkline;
